Refactor ProgramDiffModal for better reusability (split components)
